3.2-migrations-master/requirements.txt
yandex-taxi-testsuite[postgresql-binary]
